refactor: domain 과 repository 패키지 분리
